Move tx_raw_ft_transfer to simulation test
