+++ tools/integration-test/src/lib.rs
@@ -1,2 +1,3 @@
+#[allow(clippy::too_many_arguments)]
 #[cfg(test)]
 pub mod tests;
