+++ tools/test-framework/src/prelude.rs
@@ -45,6 +45,7 @@ pub use crate::framework::nary::connection::{
 pub use crate::framework::nary::node::{run_nary_node_test, NaryNodeTest, RunNaryNodeTest};
 pub use crate::framework::overrides::TestOverrides;
 pub use crate::framework::supervisor::RunWithSupervisor;
+pub use crate::ibc::denom::Denom;
 pub use crate::relayer::channel::TaggedChannelEndExt;
 pub use crate::relayer::connection::{TaggedConnectionEndExt, TaggedConnectionExt};
 pub use crate::relayer::driver::RelayerDriver;
